+++ tiskitpy/util_functions/__init__.py
@@ -4,6 +4,7 @@ from .stream_synchronize import stream_synchronize
 from .stream_unmask import stream_unmask
 from .stream_dtype import get_stream_dtype, set_stream_dtype
 from .Peterson_noise_model import Peterson_noise_model
+from .Pressure_noise_model import Pressure_noise_model
 from .coherence_significance_level import coherence_significance_level
 from .get_full_id import get_full_id
 from .match_one_str import match_one_str